requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
httpx==0.26.0
aiofiles==23.2.1
jinja2==3.1.3
pydantic==1.10.14
pydantic-settings==2.1.0
python-dotenv==1.0.0
google-generativeai==0.3.2
respx==0.20.1
pytest==7.4.4
pytest-asyncio==0.21.1
